# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Pet Manager API
  version: 0.1.0
  description: |
    REST API of Pet Manager. All endpoints are under `/api`.

    **Authentication:** send `Authorization: Bearer <token>` (from `/auth/login`
    or `/auth/register`). Data endpoints also work without a token when the
    instance runs in anonymous mode (see README). `/admin/*` always requires an
    admin token, `/household/*` a real (non-anonymous) login.

    **Households:** every account and every pet belongs to one household. All
    pet data is scoped to the caller's household; pets of other households
    answer with `404 PET_NOT_FOUND`.

    **Errors** always have the shape `{ "error": { "code", "message", "details" } }`.
    `code` is stable (e.g. `AUTH_INVALID_CREDENTIALS`, `VALIDATION_FAILED`,
    `PET_NOT_FOUND`) and is translated by the client.

    Dates are `YYYY-MM-DD`, timestamps ISO 8601 with offset. Empty strings in
    optional text fields are stored as `null`.
  license:
    name: MIT
    identifier: MIT
servers:
  - url: /api
security:
  - bearer: []
...
paths:
...
  /pets:
    get:
      tags: [pets]
      summary: List pets
      parameters:
        - { name: archived, in: query, schema: { type: boolean, default: false } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  allOf:
                    - $ref: '#/components/schemas/Pet'
                    - type: object
                      properties:
                        last_weight_kg: { type: [number, 'null'] }
                        open_illnesses: { type: integer }
    post:
      tags: [pets]
      summary: Create a pet
      requestBody:
        {
          required: true,
          content: { application/json: { schema: { $ref: '#/components/schemas/PetInput' } } },
        }
      responses:
        '201':
          {
            description: Created,
            content: { application/json: { schema: { $ref: '#/components/schemas/Pet' } } },
          }
        '400': { $ref: '#/components/responses/Error' }
...
  /household:
    get:
      tags: [household]
      summary: Own household with members and invite code
      responses:
        '200':
          {
            description: OK,
            content: { application/json: { schema: { $ref: '#/components/schemas/Household' } } },
          }
        '401': { $ref: '#/components/responses/Error' }
    patch:
      tags: [household]
      summary: Rename the own household
      requestBody:
        content:
          application/json:
            schema:
              {
                type: object,
                required: [name],
                properties: { name: { type: string, maxLength: 100 } },
              }
      responses:
        '200':
          {
            description: OK,
            content: { application/json: { schema: { $ref: '#/components/schemas/Household' } } },
          }
...
  /admin/settings:
    get:
      tags: [admin]
      summary: Instance settings
      responses:
        '200':
          {
            description: OK,
            content: { application/json: { schema: { $ref: '#/components/schemas/Settings' } } },
          }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
    put:
      tags: [admin]
      summary: Update instance settings
      description: '`requireLogin: false` is rejected with `ANONYMOUS_MODE_NOT_ALLOWED` unless the server runs with `ALLOW_ANONYMOUS_MODE=true`. Anonymous visitors then see the household of the admin who sent the request.'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                requireLogin: { type: boolean }
                allowRegistration: { type: boolean }
      responses:
        '200':
          {
            description: OK,
            content: { application/json: { schema: { $ref: '#/components/schemas/Settings' } } },
          }
        '400': { $ref: '#/components/responses/Error' }
...
  /admin/users/{id}:
    parameters: [{ $ref: '#/components/parameters/id' }]
    patch:
      tags: [admin]
      summary: Change a user's role and/or move them to another household
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                role: { type: string, enum: [admin, user] }
                household_id: { type: integer }
      responses:
        '200':
          {
            description: OK,
            content: { application/json: { schema: { $ref: '#/components/schemas/User' } } },
          }
    delete:
      tags: [admin]
      summary: Delete a user (not yourself)
      responses:
        '204': { description: Deleted }
  /admin/households:
    get:
      tags: [admin]
      summary: List all households with member and pet counts
      responses:
        '200':
          {
            description: OK,
            content:
              {
                application/json:
                  {
                    schema: { type: array, items: { $ref: '#/components/schemas/AdminHousehold' } },
                  },
              },
          }
    post:
      tags: [admin]
      summary: Create an empty household (with an invite code)
      requestBody:
        content:
          application/json:
            schema:
              {
                type: object,
                required: [name],
                properties: { name: { type: string, maxLength: 100 } },
              }
      responses:
        '201':
          {
            description: Created,
            content:
              { application/json: { schema: { $ref: '#/components/schemas/AdminHousehold' } } },
          }
  /admin/households/{id}:
    parameters: [{ $ref: '#/components/parameters/id' }]
    patch:
      tags: [admin]
      summary: Rename a household
      requestBody:
        content:
          application/json:
            schema:
              {
                type: object,
                required: [name],
                properties: { name: { type: string, maxLength: 100 } },
              }
      responses:
        '200':
          {
            description: OK,
            content:
              { application/json: { schema: { $ref: '#/components/schemas/AdminHousehold' } } },
          }
    delete:
      tags: [admin]
      summary: Delete a household without members, including all of its pets and records
      responses:
        '204': { description: Deleted }
        '409': { $ref: '#/components/responses/Error' }
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    id: { name: id, in: path, required: true, schema: { type: integer } }
...
  responses:
...
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Locale: { type: string, enum: [en, de] }
    Error:
      type: object
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, examples: [VALIDATION_FAILED] }
            message: { type: string }
            details:
              type: array
              items:
                {
                  type: object,
                  properties:
                    { path: { type: string }, code: { type: string }, message: { type: string } },
                }
    User:
      type: object
      properties:
        id: { type: integer }
        email: { type: string }
        display_name: { type: string }
        role: { type: string, enum: [admin, user] }
        locale: { $ref: '#/components/schemas/Locale' }
        household_id: { type: integer }
    Household:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        invite_code: { type: [string, 'null'] }
        created_at: { type: string, format: date-time }
        members:
          type: array
          items:
            type: object
            properties:
              id: { type: integer }
              email: { type: string }
              display_name: { type: string }
              role: { type: string, enum: [admin, user] }
    AdminHousehold:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        invite_code: { type: [string, 'null'] }
        created_at: { type: string, format: date-time }
        member_count: { type: integer }
        pet_count: { type: integer }
    Species: { type: string, enum: [dog, cat, rabbit, rodent, bird, fish, reptile, horse, other] }
    PetInput:
      type: object
      properties:
        name: { type: string, maxLength: 100 }
        species: { $ref: '#/components/schemas/Species' }
        breed: { type: [string, 'null'] }
        sex: { type: [string, 'null'], enum: [male, female, unknown, null] }
        neutered: { type: boolean }
        birth_date: { type: [string, 'null'], format: date }
        color: { type: [string, 'null'] }
        microchip: { type: [string, 'null'] }
        notes: { type: [string, 'null'] }
        archived: { type: boolean }
    Pet:
      allOf:
        - $ref: '#/components/schemas/PetInput'
        - type: object
          properties:
            id: { type: integer }
            photo: { type: [string, 'null'], description: File name under /uploads/ }
            created_at: { type: string, format: date-time }
            updated_at: { type: string, format: date-time }
...
    Settings:
      type: object
      properties:
        requireLogin: { type: boolean }
        allowRegistration: { type: boolean }
        anonymousHouseholdId:
          { type: [integer, 'null'], description: Household anonymous visitors see }
        anonymousModeAllowed: { type: boolean }
...